server: Include '%d of %d streams healthy' in top-level context
Return a structured report from generateReport, so consumers like
createHandler can extract metadata like the number of healthy streams.
Including this in the top-level thread post gives folks an idea of how
important it is to drop into the thread for details (e.g. no need to
drop in if the threads are all healthy).

Also fix threading for 'help' responses and similar, by carrying
through the message timestamp from the triggering event [1].

[1]: https://api.slack.com/messaging/sending#threading